hip_rs::core::sys
Function hipDeviceCanAccessPeer

pub unsafe extern "C" fn hipDeviceCanAccessPeer(
    canAccessPeer: *mut c_int,
    deviceId: c_int,
    peerDeviceId: c_int,
) -> hipError_t

@} / /**

@defgroup PeerToPeer PeerToPeer Device Memory Access @{ @warning PeerToPeer support is experimental. This section describes the PeerToPeer device memory access functions of HIP runtime API. / /** @brief Determine if a device can access a peer’s memory.

@param [out] canAccessPeer Returns the peer access capability (0 or 1) @param [in] deviceId - device from where memory may be accessed. @param [in] peerDeviceId - device where memory is physically located

Returns “1” in @p canAccessPeer if the specified @p device is capable of directly accessing memory physically located on peerDevice , or “0” if not.

Returns “0” in @p canAccessPeer if deviceId == peerDeviceId, and both are valid devices : a device is not a peer of itself.

@returns #hipSuccess, @returns #hipErrorInvalidDevice if deviceId or peerDeviceId are not valid devices