hip_rs::core::sys
Function hipMemPoolImportPointer

pub unsafe extern "C" fn hipMemPoolImportPointer(
    dev_ptr: *mut *mut c_void,
    mem_pool: hipMemPool_t,
    export_data: *mut hipMemPoolPtrExportData,
) -> hipError_t

@brief Import a memory pool allocation from another process.

Returns in @p dev_ptr a pointer to the imported memory. The imported memory must not be accessed before the allocation operation completes in the exporting process. The imported memory must be freed from all importing processes before being freed in the exporting process. The pointer may be freed with @p hipFree or @p hipFreeAsync. If @p hipFreeAsync is used, the free must be completed on the importing process before the free operation on the exporting process.

@note The @p hipFreeAsync api may be used in the exporting process before the @p hipFreeAsync operation completes in its stream as long as the @p hipFreeAsync in the exporting process specifies a stream with a stream dependency on the importing process’s @p hipFreeAsync.

@param [out] dev_ptr Pointer to imported memory @param [in] mem_pool Memory pool from which to import a pointer @param [in] export_data Data specifying the memory to import

@returns #hipSuccess, #hipErrorInvalidValue, #hipErrorNotInitialized, #hipErrorOutOfMemory

@see hipMemPoolExportPointer

@warning : This API is marked as beta, meaning, while this is feature complete, it is still open to changes and may have outstanding issues.

@note This API is implemented on Linux, under development on Windows.